Fix platform-specific build errors
This commit addresses several build issues that surfaced after correcting the CI cache:

- Windows: Resolved C4244 warning (float to int conversion) in PodComponent.h by adding an explicit static_cast.
- Linux: Fixed missing juce_audio_processors.h header in test builds by explicitly linking juce::juce_audio_processors to the test target.
- macOS: Resolved linker errors for DebugUIPanel, StochasticModel, and juce::dsp::LookupTable by:
    - Adding all missing .cpp source files to the plugin's CMake target.
    - Linking the juce::juce_dsp module to the plugin target.

These changes should allow the CI builds to pass successfully on all platforms.

+++ plugin/include/Pointilsynth/UI/PodComponent.h
@@ -13,7 +13,7 @@ public:
         // Placeholder: fill with a color to see it
         g.fillAll(juce::Colours::blue.withAlpha(0.5f));
         g.setColour(juce::Colours::white);
-        g.drawRect(getLocalBounds(), 1.f);
+        g.drawRect(getLocalBounds(), static_cast<int>(1.f));
         g.drawText("Pod", getLocalBounds(), juce::Justification::centred, false);
     }
